% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/logrxAddin.R
\name{logrxaddin}
\alias{logrxaddin}
\title{Creation of the logrx RStudio Addin}
\usage{
logrxaddin()
}
\value{
MiniUI for the logrx Rstudio Addin for batch submitting R files and
generating their logs
}
\description{
\code{logrxaddin} creates the miniUI for the logrx RStudio Addin
}
