% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/logrxaddin.R
\name{logrxaddin}
\alias{logrxaddin}
\title{Creation of the logrx RStudio Addin}
\usage{
logrxaddin()
}
\value{
MiniUI for the logrx Rstudio Addin for submitting an R file to logrx to
generate a log for the file.
}
\description{
\code{logrxaddin} creates the miniUI for the logrx RStudio Addin
}
\examples{
# It is recommended to run the Addin from the RStudio Addin drop down. However,
# you can call it from the command line.
\dontrun{
logrxaddin::logrxaddin()
}

}
